Initialize main thread ID after calling ctors
Experimentation with passive segments uncovered the fact that global
constructors were being run after the main thread id was
initialized. In the C code the main thread id is initialized to zero,
so when the BSS was being loaded into memory it was clobbering the
manually initialized value. This issue only surfaces with passive
segments because the BSS is otherwise optimized out by binaryen.