Nicer errors for missing exceptions symbols that can happen due to -fno-exceptions (#8966)
With that flag at link time we don't link in any symbols for throwing, but if source files have that symbol, that will lead to undefined symbol errors. Instead, show a nice error explaining the situation and what to do.

Also, -fno-exceptions -s DISABLE_EXCEPTION_CATCHING=0 is wrong as well: if throwing is not allowed, then it is meaningless to enable exceptions. Also show a nicer error for this case.

possibly helps with #8963