CI: Disable chrome cache (#8977)
Recently we've gotten errors where one test loads a JS+wasm combo that contains either the JS or the wasm from the last test (!), manifesting as "import is not a function" as the imports provided by JS don't match the mismatched wasm. Oddly this happened on sdl_pumpevents multiple times on CI.

As this is only on Chrome, my guess is that it's a caching issue, but I can't reproduce it locally to verify. Disabling the cache seems to help, at least the run here was fully successful.