Fix cache testing on CI, and fix a regression from #8973 (#8994)
The emsdk ships libc and a few other libs, and we used those, instead of building libc here - so we missed a regression that was only caught later on releases CI. I'm surprised we didn't hit more problems earlier, actually...

The regression was that we added a constructor priority in pthreads, and apparently that keeps it not just alive but separate from other ctors, which adds an extra function to the number. I refactored that code to make it clear what we do in the pthreads and non-pthreads cases, and made us only use the priority when using pthreads.