Wasm backend LTO docs and testing (#8942)
Documents the -flto and WASM_OBJECT_FILES=0 options, which are not identical as the latter also affects system libraries. Also document that --llvm-lto 1 is needed to actually run lto opts (otherwise we pass --lto-O0 to wasm-ld).

Add a test for -flto and WASM_OBJECT_FILES=0 emitting bitcode files as expected.