Fix browser harness logging (#8980)
Turns out at least some of the duplicate logging we have is because if a python web request hander doesn't return any headers, it just calls it again. Odd that this doesn't recurse infinitely... anyhow, just send some empty headers for the logging code to avoid that.

Also fix the logging in python3, where urllib is structured differently.