Asyncify: fix abort() handling (#9000)
Asyncify wasm error handling leads to unreachables being thrown. This PR maps those errors into calls to abort(), which ensures the runtime is properly shut down if they occur. In particular, no more code will be run, which could previously be confusing while debugging.

Background: Asyncify is designed as a standalone binaryen pass - it knows nothing about emscripten. You can run it on arbitrary wasm. So it uses the universal unreachable error mechanism. (We could add an "emscripten" mode to it there, but I preferred to keep the emscripten-specific details here in emscripten.)