Fix one flakiness issue in browser.test_chunked_synchronous_xhr (#9003)
The test is marked @flaky which runs it up to 3 times to try to avoid a random error. But we never shut down the server, so the second and later runs would just hit an error on trying to create a server on the same port. (Of course not shutting down the server was bad for other reasons too!)

Hopefully this will decrease dramatically the number of random failures on this test, but it won't eliminate them entirely.