Make STACK_BASE refer to the starting stack location on threads
This is already a case on the main thread. However, on WASM backend,
STACK_BASE == STACK_MAX == the lowest address on the stack, i.e. the
point where the stack can grow to. This makes it impossible to find
the highest address on the stack, i.e. the bottom of the stack.

This PR makes the behaviour consistent between the main thread and
other threads by making other threads follow the conventions of the
main thread.