Make read, readAsync, readBinary, setWindowTitle normal JS variables (#9047)
This replaces Module['readBinary'] with readBinary etc., that is, moves them from Module to be just normal JS variables.

This change allows better JS minification: the names read, readAsync etc. can be minified now, and even better if the variables are not used then they can be removed entirely. On hello world this saves 4% of JS size in -O3 --closure 1 with the wasm backend.

This is an internal API change. However, it's possible some users depend on it, so I added a mention in the changelog, and in builds with ASSERTIONS an error message is shown if the old APIs are used.

Note that read is also renamed to read_ (since "read" is an API call in the SpiderMonkey shell).