Fix #9068: get rid of malformed newlines in emcc errors on Windows
This fixes a patch applied to acorn.js for quoting error-causing text
in parse errors (https://github.com/acornjs/acorn/pull/793).
Previously it splitted binary data on \n to get lines, which was
inconsistent with the rest of the parser and caused excess \r in
the error message on Windows. Now it uses the same regex as the rest
of acorn.js (e.g. getLineInfo used by error reporting).