Asyncify whitelist and blacklist support (#9094)
A slight complication here is that we need function names for the lists. This extends the current symbol-map mechanism to keep around function names in the wasm until the very end of the optimization pipeline. (There are still risks with inlining and duplicate function elimination changing things, though, which can't be solved without adding some new special function annotation.)

Uses WebAssembly/binaryen#2218