More INCOMING_MODULE_JS_API optimizations (#9084)
Module.{wasmBinary,wasmMemory,TOTAL_MEMORY,print,printErr} are optimized for that option. This should not be user-observable (as by default we assume they might be used; but if the option is used to say they will not, then we can optimize them out).

Reduces ./emcc tests/hello_world.c -O3 --closure 1 -g1 -s 'INCOMING_MODULE_JS_API=[]' -s ENVIRONMENT=web by 5%.

Added a helper makeModuleReceiveWithVar that defines var, and gets a value from Module for it, if we expect one. This cleans up a bunch of code.