Parse -s changes using the json parser where possible, allowing human-readable C++ names in -s lists, which is important for Asyncify (#9132)
We have nice syntax for lists, -s LIST=[foo,bar] which works
well in many cases. But sometimes the list contains items
that contain ",", for example -s LIST=["foo(x, y)", "bar"]. To handle
that properly, try to parse things that look like json using the json
parser; if it fails go back to the old code.

Together with https://github.com/WebAssembly/binaryen/pull/2275 
this fixes #9128 where the lists there actually do have ","s in them.

Add testing for Asyncify lists with such names, and also added
extra warnings for when escaping is confusing, and a test for that
as well.