emscripten_scan_registers and emscripten_scan_stack (#9121)
These functions can help conservative GCs.

emscripten_scan_stack provides an easy way to find where the C stack is in memory.

emscripten_scan_registers allows scanning "registers" - things *not* in memory. In wasm that means local vars, including in functions higher up the stack. This uses Asyncify in order to be able to spill all the local state and unwind the stack, at which point it is possible to scan the data, then we rewind and resume. This does incur Asyncify's overhead, so unlike emscripten_scan_stack it has a downside.