wasm2js testing (#9226)
This adds wasm2js1 to CI here. This is the first large amount of wasm2js testing we add. I picked -O1 because it's a mode we don't test for upstream currently, so this adds indirect coverage for that, and also -O1 tests quite a bit of optimizations in wasm2js.

To balance CI load, I removed some old asm.js modes from testing, asm1 (asm.js with -O1) and asmi (asm.js with emterpreter at -O0). We do still test asm.js at -O0, -O2 (and some in other) so removing -O1 seems not a big risk given fastcomp is going away, and the emterpreter stuff is still tested at -O2 (which itself is perhaps excessive, but I didn't want to remove too much at once).

This removes the scary warning that wasm2js is not ready for normal use. After the initial bugreports we've not gotten any for a while, so I think it's usable.