Don't build embind as a library
This is partial revert of #8817.  It turns out that all users of bind.h
have to agree on -fno-rtti.  In particular if a user passed -fno-rtti
to build their code the bind.cpp needs to be compiled with the same
settings.

Perhaps we can find a way to avoid this in the future but for now
it looks like we are stuck compiling bind.cpp on demand at link time.

Fixes: #9122