Include debug info in system libraries
Add `-g4` when building system libraries.  This should effect the
optimization level but allows users who link with `-g4` themselves
to have symbol information for system libraries as well as their
own code.

As far as I can tell the only downside to doing this is that slight
increase in the on-disk side of the system libraries and the minor cost
of the linker have to read this info.