Honor SIDE_MODULE=2 with wasm backend
With SIDE_MODULE=2 we don't want to export all visibility `default`
symbols.  Only the symbols we explicitly export from the command line.

For this to actually work we need to wait for the following change
to land in llvm: https://reviews.llvm.org/D66359

This change is part of the fix for #9038.