Change test suite do_run default to checking the return code (#9252)
Before this, a new test could still pass if running the code ends with an error - we check the output text, which is usually enough. But seems safer to require tests that can exit with an error to explicitly state that.

Also clean up core_test_modes which should be different on the wasm backend, and remove old assertions on test modes existing in fastcomp etc. (they won't exist for long anyhow, and make debugging less convenient, e.g., they make tests/runner.py randomX not work).