Add error handling in ASSERTIONS for EM_JS errors with ASYNCIFY_IMPORTS (#9358)
This matches what we already have for js library functions.

This also disables wasm/JS import minification when ASSERTIONS are on. That makes this much simpler to write (no need to worry about names changing for these assertion checks), and really it seems better anyhow - in ASSERTIONS code size is not crucial, and that's the only purpose for that minification. And it makes debugging such builds easier, which is usually the point of ASSERTIONS.

Also add more tests, including response file stuff.