Support symbols files for different options (#9433)
This aims at automating symbol generation a bit more and supporting
different symbols files for different set of build options. For example,
a library can be built with different set of options: such as with `-mt`
and `-noexcept`, only `-mt`, only `-noexcept`, or with none of them. And
libraries built with those different options can expose different set of
symbols. Currently this kind of info is hard-coded into the code.

After this PR, all system libraries read symbols file in
`system/lib/symbols/wasm/library_filename.symbols` when using the wasm
backend, and `system/lib/symbols/asmjs/library_filename.symbols` when
using fastcomp. For example, if the filename of a library we are
considering to link is `libc++-mt.a`, it will search for
`system/lib/symbols/wasm/libc++-mt.symbols`. And if it doesn't exist, it
will try the default library name without any options:
`system/lib/symbols/wasm/libc++.symbols`.

`tools/update_symbols.py` generates all symbols files automatically. It
does not have `--asmjs` anymore: it will read `.emscripten` just as the
normal emcc toolchain does. This script has `target_libs` variable, in
which we only list several libraries that need symbol files. (This set
of libraries is the libraries we currently have `.symbols` files.)
This script builds all possible combination of those libraries if they
don't exist in cache, and run `llvm-nm` on them, and generate files
within `system/lib/symbols/wasm` or `system/lib/symbols/asmjs`
directories. As mentioned above, to prevent proliferation of symbols
files with same contents, if symbols files generated with different
options are the same as the 'default' symbol file without any suffices,
those symbol files will be deleted and the linking algorithm will fall
back to the default symbol files for the library.

One of the reasons I did this was, when we add the new wasm exception
handling, the set of symbols exposed by different options for libraries
such as libcxxabi will certainly be different, and I hoped to make it
more general.