Support synchronous execution "concurrently" with ASYNCIFY asynchronous execution (#9412) (#9423)
Makes changes related to discussion in #9412.

Changed Behavior:

    Supports synchronous execution "concurrently" with ASYNCIFY asynchronous execution (as per discussion in #9412)
    Slight downside is due to the re-factor, we can only detect and warn against two in-flight async calls if they are properly labelled by {async: true} in opts passed to ccall/cwrap. If the user doesn't properly mark {async: true} and runs an async-ified function concurrently while an in-flight async call is happening, it will return an undefined value "synchronously" but run concurrently with the other in-flight async call leading to likely memory corruption.

The reason for this downside is before, we would basically say any call to ccall/cwrap while an async function was in-flight should not be allowed. Now that we are allowing synchronous ccall/cwraps while an async function is in-flight, we can only detect multiple in-flight async calls if they are properly labelled with {async: true} in opts.

The downside seems fine to me as the user would already be aware of the {async: true} option to even get into this state (as one asynchronous call has to be in-flight already), so it shouldn't hurt new users who aren't aware of that option existing and the likely quick memory corruption and failure would signal to them the second async call needs to be fixed by adding the proper opts. But just something to note.