Calculate wasm "env" imports at compile time (#9447)
This moves that code from runtime addition of fields to the env, to compile-time addition and emitting of the final import object as a whole.

This has no significant functional change. It just makes the emitted code smaller and simpler. This also simplifies the overall model, and will allow further refactoring to simplify the dynamic linking code (which currently has its own computation of the wasm imports), leaving that for a later PR to keep this one as small as possible. This will also help with the code splitting experiment in the future.

This fixes the acorn parser passes to handle wasm import fields that are not just names, as now they can also be numbers or more complex expressions (the stuff we compute at compile time now may include such things).

The metadce tests get lots of updates, but there is no substance to them: the tests look at the initial wasm import object, before we added more fields at runtime. We don't add the fields at runtime anymore, so the tests can see those extra fields now, but they were always there, just the tests didn't notice them before. So there is no increase in the number of wasm imports, it's the same.

When testing this I also noticed running random tests was broken on fastcomp, we were just missing an if in runner.py, I added that here.