Fix test_minimal_dynamic with wasm backend (#9454)
There were two issues here. Firstly, printf was being lowered to iprintf
in the side module.  Secondly we were mapping EXPORT_ALL emscripten
option to `--export-all` linker flag which was doing the wrong thing.
`--export-all` in the linker basically prevents GC of all symbols but
the EXPORT_ALL emscripten option is not supposed to do that.

Fixes: #9038