Perform "linking to object file" as separate phase. (#9444)
This change moves the "linking to object file" logic to separate
phase which runs after compilation, but before the real "linking"
phase.  We can also delay library processing until after this phase
as that is only needed for actual linking.

This change is a step towards more conventional library linking
See #8912.