Limit object file linking to certain file extensions.
Previously if the output file didn't have an extension we would
outout an object file.  This might have been useful for tools like
autoconf that output binaries with no extension by default but its
somewhat confusing to get an object file without actually asking for
one.  So now we default of js output.