Add test that allocates 50 threads 10x times. Use the default parameter set for preallocated threads. I checked this with my change and without my change.
With my change, the initial allocation takes ~27.66 ms.
Without my change, the initial allocation takes 28.67 ms.

As we are talking about a delta of ~1ms, this isn't much of a difference.

Note that this is the version of the test that checks that there are no unexpected changes when the defaults are used when allocating threads. I'm not sure how to measure startup time in a test setting...