Fix JS memory handling for STANDALONE_WASM (#9474)
Don't create the wasm Memory in JS since wasm will do so in standalone mode. With that done, various things need fixing since they assumed memory was already present, which is the rest of the work in this PR.

Move the memory init code into a new file, runtime_init_memory.js (no changes to that code). The file is conditionally included if not standalone.

Fix stack cookie checking for standalone mode:

*    Write the cookie after the wasm memory actually arrives.
*    Do the endianness check in a new buffer (it's just in assertions mode, so perf isn't an issue, and this disentangles it from the main memory, which in standalone mode isn't ready yet).
*    Move the address-0-integrity write into writeStackCookie, which matches where it is checked, which is in checkStackCookie. That way it is done at the right time and not synchronously during startup (which is invalid in standalone mode).

Add more testing of standalone wasm in -O0, now that this stuff works.