Update intArrayToString to match docs, and update docs
intArrayToString is documented as taking null-terminated input,
but previously did not stop at a 0 entry.

Switched it to use UTF8ArrayToString internally using the
array length as maximum length. This both fixes the discrepancy
with the documentation by allowing 0-terminated input arrays,
and improves general compatibility by allowing input data to
contain UTF-8 strings without being corrupted on output to
JavaScript.

Updated documentation for both intArrayToString and its
almost-inverse, intArrayFromString, to indicate that UTF-8
is used (this was already the case, but not documented,
for intArrayFromString).

Fixes #4840