Always run --remove-unused-module-elements when STANDALONE_WASM (#9507)
We always run that pass when optimizing, but normally do not when not
optimizing. In standalone mode, though, an unnecessary import may make the wasm
not standalone and so not runnable in wasmer/wasmtime/etc. So in that mode, it's
best to run it, to avoid a silly import breaking the wasm.

A concrete example of when this is needed is that in hello world with iostreams
we end up with unnecessary imports of atexit (we will need to fix that
eventually, but it's only needed when EXIT_RUNTIME is flipped by the user, as
the default ignores atexits).