Pass `-c` when compiling SDL2 object files
Without `-c` emscripten will try to run lld on the object files its
given to produce another object file.  We should probably disable that
for the single build case.  This breakage was introduced in #9444 and
we should probably address it, but compiling without `-c` is also
broken so fixing here.

Fixes #9510