Proper STANDALONE_WASM testing, using WASM_ENGINES (#9499)
The initial landing of STANDALONE_WASM support had testing, but it was pretty
hackish, just enough to get us going. This adds more proper support.

This defines WASM_ENGINES in the settings file, parallel to JS_ENGINES. Adds
support for running in those runtimes - in particular, we run the wasm and not
the JS.

Adds also_with_standalone_wasm to core, which allows us to add files to be
tested in this mode. So far I added two tests.

Ignore temp file leaks on /tmp/wasmer/ as it appears to add a bunch of temp
files there sometimes.