Use wasi for getenv etc. (#9525)
This uses musl's getenv etc. implementations for the wasm backend, and
initializes the data using the wasi APIs.

This doesn't change the fastcomp implementation. There seems to be no good way
to do it, as libc is bitcode, so the global ctor for this would mean it's always
 included, and we can't put it in libc-extras since we don't know when it's
needed (e.g. libc++ might use getenv, but we don't track such internal deps,
and adding hacks to the linker seems risky). So fastcomp still uses the old
impl in JS.

In the new upstream impl, the wasi functions call emscripten_get_environ which
sets up the default env values, which are the same as before.

This makes library_wasi.js always enabled, and not conditionally.

This fixes a bug in our current getenv impl, which led to a test change. I also
removed the test for unsetenv(0) which is an invalid use of the API and which
musl does not handle.