Add a default for JS_ENGINES
Since we are guaranteed that `NODE_JS` is part of the config (since
all users need this) we can set a sensible default for `JS_ENGINES`.

This default was supposed to be part of #9469, but a bug meant the code
was dead and it was removed in #9499.

`JS_ENGINES` is needed for any users that want to run test code.

Fix `gen_struct_info.py` to explicitly use NODE_JS since this is the
only engine we expect users to have installed and the only one we
officially support outside of test code.