Rename errno constants to use the wasi values (#9545)
Redefine errno values to be consistent with wasi. This will let us avoid
needing to convert the values back and forth as we use more wasi APIs (which
I experiemented with and it adds 1K or so).

This is an ABI change, which should not be noticeable from user code
unless you use errno defines (like EAGAIN) and keep around binaries
compiled with an older version that you link against. In that case, you
should rebuild them.

Fix NODEFS, which basically just routed the node errno code to ours, which
only worked when the underlying system had codes similar to musl - which seems
to have been the case on some linuxes. This does make NODEFS larger, as it
uses the ERRNO lookup table now, I'll look into a way to avoid including
it unnecessarily as a followup (would be a breaking change).

Most of the changes in this PR are test changes, places where we printed
raw errno codes.