Replace musl close syscall with wasi fd_close (#9552)
Musl has a bunch of such calls scattered around, so this required changes to
more than 1 or 2 places.

This also adds basic support for handling wasi syscall return values, which are
different than musl's (which are negated).

This includes the disabling of ASMFS from #9551. If we don't decide to to land
that, we'd need to do more work to port ASMFS to this new syscall.

This enables -Wno-unused-result on musl, as musl often calls close and ignores
the result, and the wasi syscalls are marked so that they warn on that.
Alternatively I can modify the wasi header? Or is there a workaround in the call
site to say "ignore the value"?

Separately, add stubs for mmap syscalls (I can split that out, but it's 6 lines,
so seems silly, and it's next on my list).