Temporarily change throw() to noexcept for wasm EH
Clang frontend for Wasm EH does not yet support exception specification,
such as `throw()` or `throw(type)`. It supports `noexcept` fine.
Currently all uses of exception specification in libcxxabi are
`throw()`, so before going all the way down to support exception
specification in clang, which will take a while, this converts them to
`noexcept` for wasm EH. `throw()` and `noexcept` are not exactly
identical in their semantics: `throw()` calls `std::unexpected()` when
one of its instructions throws and `noexcept` calls `std::terminate()`
in the same case. But I don't think it will cause visible difference
from the user side.

This does not have a test case because this should be tested with wasm
EH support, which will be uploaded later. The reason I made this as a
separate PR is, unlike other changes for wasm EH, this is meant to be
temporary, and making this a separate PR makes it easy to revert.