Remove EMCONFIGURE_JS (#9603)
This finishes the long-term process of getting our configure test support to
always use JS, and not depend on building natively.

Since #6269 that flag was "2" which means never use native, always use JS,
and there have not been many bug reports about it, so it seems to work.

This may break people that used the env var to override the default, however,
it wasn't very discoverable in the docs (had to read emcc.py) so hopefully no
one will be affected. If bug reports come in, we should just fix what is
preventing configure from running in JS for them (i.e. I don't think we'd ever
want to bring this env var back).